Split gem_model into print_table and score_methods
